partial refactoring into component folders, though containers not really separated accurately
